Move zoom operation to ViewPortTransform module (#31)
