+++ boids.cabal
@@ -20,6 +20,7 @@ library
                      , xml
   exposed-modules:     Boids
                      , TestData
+                     , ViewPortTransform
   default-language:    Haskell2010
 
 executable boids-exe
