+++ test/BoidsSpec.hs
@@ -1,4 +1,4 @@
-module BoidsSpec (testGuideBoidToAngle, testZoom) where
+module BoidsSpec (testGuideBoidToAngle) where
 
 import Test.HUnit
 import Test.HUnit.Approx
@@ -25,18 +25,3 @@ testGuideBoidToAngle =
                                   }
                   guidedBoid = guideBoidToAngle angle testBoid
               in TestCase (assertApproxEqual message errorMargin expectedSteer $ boidSteer guidedBoid)
-
-testZoom :: Test
-testZoom = TestList [ TestLabel "Zoom In" zoomInTestCase
-                    , TestLabel "Zoom Out" zoomOutTestCase
-                    , TestLabel "Zoom Out Saturation" zoomOutSaturationTestCase
-                    ]
-    where zoomInTestCase = generalTestCase 1.0 0.5 1.5
-          zoomOutTestCase = generalTestCase 1.0 (-0.5) 0.5
-          zoomOutSaturationTestCase = generalTestCase 0.1 (-0.5) 0.0
-
-          generalTestCase initialScale zoomStep expectedScale =
-              let viewPort = viewPortInit { viewPortScale = initialScale }
-                  zoomedScale = viewPortScale $ zoom zoomStep viewPort
-                  message = "Unexpected result scale"
-              in TestCase (assertApproxEqual message errorMargin expectedScale zoomedScale)
